refactor: implemented a huge refactor as a part of performance optimization and UI enhancements process. instead of showing the add todo component within the home page, now I moved it within a modal. now add todo component wouldn't render untill the user has clicked on add todo button. also improved UI of todo lists component. refactor process is still ongoing.
