+++ src/components/todo-container/TodoList.tsx
@@ -1,35 +1,25 @@
-import { usePagination } from '@/hooks/usePagination';
-import { useSearch } from '@/hooks/useSearch';
-import { currentSelectedUserId } from '@/redux/features/auth/auth-slice/authSlice';
-import {
-	useDeleteTodoMutation,
-	useGetCategoriesQuery,
-	useGetTodosQuery,
-	useUpdateTodoMutation,
-} from '@/redux/features/todo/todo.api';
-import { useAppSelector } from '@/redux/hooks';
+import { useDeleteTodoMutation, useGetCategoriesQuery, useUpdateTodoMutation } from '@/redux/features/todo/todo.api';
 import type { TodoType } from '@/types';
 import { getCategoryById } from '@/utils/getCategoryById';
 import { getPriorityColor } from '@/utils/getPriorityColor';
 import { createMarkdownContent } from '@/utils/markdown-utils/createMarkdownContent';
 import MDEditor from '@uiw/react-md-editor';
-import { Eye, EyeClosed, PenBox, Trash } from 'lucide-react';
+import { Calendar, CheckCircle2, Clock, Eye, EyeClosed, PenBox, Tag, Trash } from 'lucide-react';
 import { lazy, Suspense, useState } from 'react';
 import { toast } from 'sonner';
-import PaginationButtons from '../pagination-buttons/PaginationButtons';
 import { Badge } from '../ui/badge';
 import { Button } from '../ui/button';
-import { Checkbox } from '../ui/checkbox';
 import { ScrollArea } from '../ui/scroll-area';
-// import EditModal from '../edit-modal/EditModal';
+
 const EditModal = lazy(() => import('@/components/edit-modal/EditModal'));
 const DeleteModal = lazy(() => import('@/components/delete-modal/DeleteModal'));
 
-const TodoList = ({ searchQuery }: { searchQuery: string }) => {
-	const userId = useAppSelector(currentSelectedUserId);
-
-	// getting todos and categories using rtk query hooks
-	const { data: todosData } = useGetTodosQuery({ userId });
+const TodoList = ({
+	searchQuery,
+	currentItems,
+	filteredTodosAfterSearch,
+}:{searchQuery:string;currentItems:TodoType[];filteredTodosAfterSearch:TodoType[]}) => {
+	// getting categories using rtk query hooks
 	const { data: categoriesData } = useGetCategoriesQuery();
 	const [updateTodo] = useUpdateTodoMutation();
 	const [deleteTodo] = useDeleteTodoMutation();
@@ -65,10 +55,7 @@ const TodoList = ({ searchQuery }: { searchQuery: string }) => {
 			});
 	};
 
-	// edit todo related logc end
-
 	// delete todo related logic start
-	// handle delete click handler
 	const handleDeleteClick = (todo: TodoType) => {
 		setSelectedTodo(todo);
 		setIsModalOpen(true);
@@ -89,8 +76,6 @@ const TodoList = ({ searchQuery }: { searchQuery: string }) => {
 		setSelectedTodo(null);
 	};
 
-	// delete todo related logic end
-
 	// toggle todo status handler
 	const handleTodoToggle = async (id: string, currentStatus: boolean) => {
 		try {
@@ -108,194 +93,236 @@ const TodoList = ({ searchQuery }: { searchQuery: string }) => {
 	// expand toggle function
 	const expandTodoToggle = (id: string) => {
 		if (expandedTodoId === id) {
-			setExpandedTodoId(null); // close if already expanded
+			setExpandedTodoId(null);
 		} else {
-			setExpandedTodoId(id); // otherwise expand the todo
+			setExpandedTodoId(id);
 		}
 	};
 
-	// pagination related logic starts
-	const filteredTodosAfterSearch = useSearch(todosData?.todos || [], searchQuery, [
-		'title',
-		'description',
-		'priority',
-		'category',
-	]);
-
-	const {
-		currentPage,
-		totalPages,
-		currentItems,
-		goToSpecificPage,
-		goToNextPage,
-		goToPreviousPage,
-		goToFirstPage,
-		goToLastPage,
-	} = usePagination({ items: filteredTodosAfterSearch, itemsPerPage: 2 });
-
 	return (
-		<ScrollArea className='h-[500px] rounded-lg border border-zinc-800 bg-zinc-900 p-4'>
-			<div className='space-y-4'>
-				{/* Empty state */}
-				{filteredTodosAfterSearch.length === 0 && (
-					<div className='text-center text-zinc-400 py-12 rounded-lg border border-dashed border-zinc-700'>
-						<p className='text-sm font-medium'>No todos found</p>
-						<p className='text-xs mt-1 text-zinc-500'>Start by creating your first todo!</p>
-					</div>
-				)}
+		<div className='space-y-4'>
+			<ScrollArea className='h-[600px] pr-4'>
+				<div className='space-y-4'>
+					{/* Empty state */}
+					{filteredTodosAfterSearch.length === 0 && (
+						<div className='flex flex-col items-center justify-center py-16 text-center'>
+							<div className='w-16 h-16 rounded-full bg-zinc-800 flex items-center justify-center mb-4'>
+								<CheckCircle2 className='w-8 h-8 text-zinc-600' />
+							</div>
+							<h3 className='text-lg font-semibold text-zinc-300 mb-2'>No todos found</h3>
+							<p className='text-sm text-zinc-500 max-w-md'>
+								{searchQuery
+									? `No todos match "${searchQuery}". Try adjusting your search.`
+									: 'Start by creating your first todo to get organized!'}
+							</p>
+						</div>
+					)}
 
-				{/* Todo items */}
-				{currentItems.map((todo: TodoType) => {
-					const isCompleted = todo.completed === true;
-					const category = getCategoryById(todo.category, categoriesData!);
-					// expanded content
-					const isExpanded = expandedTodoId === todo?.id;
-					const markdownTodoContent = createMarkdownContent(todo?.title || '', todo?.description || '');
-					const previewContent = markdownTodoContent.split('\n').slice(0, 2).join('\n');
+					{/* Todo items */}
+					{currentItems.map((todo: TodoType) => {
+						const isCompleted = todo.completed === true;
+						const category = getCategoryById(todo.category, categoriesData!);
+						const isExpanded = expandedTodoId === todo?.id;
+						const markdownTodoContent = createMarkdownContent(todo?.title || '', todo?.description || '');
+						const previewContent = markdownTodoContent.split('\n').slice(0, 2).join('\n');
 
-					return (
-						<div
-							key={todo.id}
-							className='flex items-start justify-between rounded-xl border border-zinc-700 cursor-pointer bg-zinc-800/50 p-4 shadow-sm hover:border-zinc-600 transition-colors'>
-							{/* Left side */}
-							<div className='flex gap-3 flex-1'>
-								{/* Checkbox */}
-								<Checkbox
-									checked={isCompleted}
-									onCheckedChange={() => handleTodoToggle(todo.id as string, isCompleted)}
-									className='mt-1'
-									aria-label='Toggle for todo completed and uncompleted'
-								/>
+						return (
+							<div
+								key={todo.id}
+								className={`group relative rounded-xl border transition-all duration-300 ease-in-out ${
+									isCompleted
+										? 'border-zinc-700/50 bg-zinc-800/30'
+										: 'border-zinc-700 bg-zinc-800/50 hover:border-zinc-600 hover:bg-zinc-800/70'
+								} ${isExpanded ? 'shadow-lg' : 'shadow-sm hover:shadow-md'}`}>
+								{/* Priority indicator */}
+								{todo.priority && !isCompleted && (
+									<div
+										className={`absolute left-0 top-0 bottom-0 w-1 rounded-l-xl ${
+											todo.priority === 'high'
+												? 'bg-red-500'
+												: todo.priority === 'medium'
+													? 'bg-yellow-500'
+													: 'bg-green-500'
+										}`}
+									/>
+								)}
 
-								{/* Content */}
-								<div className='flex-1 space-y-5'>
-									{/* Markdown content */}
-									<div className={isCompleted ? 'line-through opacity-60' : ''}>
-										<MDEditor.Markdown
-											source={isExpanded ? markdownTodoContent : previewContent}
-											style={{
-												backgroundColor: 'transparent',
-												color: isCompleted ? '#a1a1aa' : '#e4e4e7',
-												fontSize: 14,
-												lineHeight: 1.6,
-											}}
-										/>
+								<div className={`p-6 transition-all duration-300 ${isExpanded ? 'pb-8' : ''}`}>
+									{/* Header */}
+									<div className='flex items-start gap-4'>
+										{/* Custom Checkbox */}
+										<button
+											type='button'
+											onClick={() => handleTodoToggle(todo.id as string, isCompleted)}
+											className={`mt-1 flex-shrink-0 w-5 h-5 rounded-full border-2 flex items-center justify-center transition-all duration-200 ${
+												isCompleted
+													? 'border-green-500 bg-green-500 text-white'
+													: 'border-zinc-600 hover:border-zinc-500 group-hover:border-zinc-400'
+											}`}
+											aria-label='Toggle todo completion'>
+											{isCompleted && <CheckCircle2 className='w-3 h-3' />}
+										</button>
+
+										{/* Content */}
+										<div className='flex-1 min-w-0 space-y-3'>
+											{/* Markdown content */}
+											<div className={`transition-all duration-200 ${isCompleted ? 'opacity-60' : ''}`}>
+												<div className={isCompleted ? 'line-through' : ''}>
+													<MDEditor.Markdown
+														source={isExpanded ? markdownTodoContent : previewContent}
+														style={{
+															backgroundColor: 'transparent',
+															color: isCompleted ? '#a1a1aa' : '#e4e4e7',
+															fontSize: '14px',
+															lineHeight: '1.6',
+															fontFamily: 'inherit',
+														}}
+													/>
+												</div>
+											</div>
+
+											{/* Metadata */}
+											<div className='flex flex-wrap gap-2'>
+												{/* Due Date */}
+												{todo.dueDate && (
+													<Badge variant='secondary' className='bg-amber-500/20 text-amber-300 border-0 font-medium'>
+														<Calendar className='w-3 h-3 mr-1' />
+														Due: {new Date(todo.dueDate).toLocaleDateString()}
+													</Badge>
+												)}
+
+												{/* Priority */}
+												{todo.priority && (
+													<Badge
+														variant='secondary'
+														className={`${getPriorityColor(todo.priority)} border-0 font-medium`}>
+														<div
+															className={`w-2 h-2 rounded-full mr-2 ${
+																todo.priority === 'high'
+																	? 'bg-red-500'
+																	: todo.priority === 'medium'
+																		? 'bg-yellow-500'
+																		: 'bg-green-500'
+															}`}
+														/>
+														{todo.priority.charAt(0).toUpperCase() + todo.priority.slice(1)}
+													</Badge>
+												)}
+
+												{/* Category */}
+												{category && (
+													<Badge variant='secondary' className='bg-zinc-700/60 text-zinc-300 border-0'>
+														<Tag className='w-3 h-3 mr-1' />
+														<div className='w-2 h-2 rounded-full mr-1' style={{ backgroundColor: category.color }} />
+														{category.name}
+													</Badge>
+												)}
+
+												{/* Created Date */}
+												<Badge variant='secondary' className='bg-zinc-700/40 text-zinc-400 border-0'>
+													<Clock className='w-3 h-3 mr-1' />
+													Created {new Date(todo.createdAt as string).toLocaleDateString()}
+												</Badge>
+
+												{/* Status */}
+												<Badge
+													variant='secondary'
+													className={`border-0 font-medium ${
+														!isCompleted ? 'bg-amber-500/20 text-amber-400' : 'bg-green-500/20 text-green-400'
+													}`}>
+													{!isCompleted ? 'Pending' : 'Completed'}
+												</Badge>
+											</div>
+										</div>
+
+										{/* Actions */}
+										<div className='flex-shrink-0 flex gap-2'>
+											<Button
+												onClick={() => expandTodoToggle(todo.id)}
+												size='sm'
+												variant='ghost'
+												className='h-8 px-3 text-zinc-400 hover:text-zinc-200 hover:bg-zinc-700/60'
+												aria-label={isExpanded ? 'Collapse todo' : 'Expand todo'}>
+												{isExpanded ? <EyeClosed className='w-4 h-4' /> : <Eye className='w-4 h-4' />}
+											</Button>
+
+											<Button
+												onClick={() => handleEditClick(todo)}
+												size='sm'
+												variant='ghost'
+												className='h-8 px-3 text-zinc-400 hover:text-zinc-200 hover:bg-zinc-700/60'
+												aria-label='Edit todo'>
+												<PenBox className='w-4 h-4' />
+											</Button>
+
+											<Button
+												onClick={() => handleDeleteClick(todo)}
+												size='sm'
+												variant='ghost'
+												className='h-8 px-3 text-zinc-400 hover:text-red-400 hover:bg-red-500/10'
+												aria-label='Delete todo'>
+												<Trash className='w-4 h-4' />
+											</Button>
+										</div>
 									</div>
 
-									{/* Metadata badges */}
-									<div className='flex flex-wrap gap-2'>
-										{/* Created */}
-										<Badge variant='secondary' className='bg-zinc-700/50 text-zinc-300'>
-											Created: {new Date(todo.createdAt as string).toLocaleDateString()}
-										</Badge>
-
-										{/* Due Date */}
-										{todo.dueDate && (
-											<Badge variant='secondary' className='bg-amber-500/20 text-amber-300'>
-												Due: {new Date(todo.dueDate).toLocaleDateString()}
-											</Badge>
-										)}
-
-										{/* Priority */}
-										{todo.priority && (
-											<Badge variant='secondary' className={`${getPriorityColor(todo.priority)} bg-zinc-700/40`}>
-												{todo.priority.charAt(0).toUpperCase() + todo.priority.slice(1)}
-											</Badge>
-										)}
-
-										{/* Category */}
-										{category && (
-											<Badge variant='secondary' className='bg-zinc-700/50 text-zinc-300 flex items-center gap-1'>
-												<div
-													className='w-2 h-2 rounded-full border border-zinc-600'
-													style={{ backgroundColor: category.color }}
-												/>
-												{category.name}
-											</Badge>
-										)}
-
-										{/* Status */}
-										<Badge
-											variant='secondary'
-											className={`${
-												!isCompleted ? 'bg-amber-500/20 text-amber-400' : 'bg-green-500/20 text-green-400'
-											}`}>
-											{!isCompleted ? 'Pending' : 'Completed'}
-										</Badge>
+									{/* Status indicator at bottom */}
+									<div
+										className={`mt-4 pt-4 border-t border-zinc-700/50 transition-all duration-300 ${
+											isExpanded ? 'opacity-100' : 'opacity-0 h-0 mt-0 pt-0 border-t-0 overflow-hidden'
+										}`}>
+										<div className='flex items-center justify-between text-xs'>
+											<div className='flex items-center gap-4 text-zinc-500'>
+												<span>Status: {isCompleted ? 'Completed' : 'In Progress'}</span>
+												{todo.dueDate && <span>Due: {new Date(todo.dueDate).toLocaleDateString()}</span>}
+											</div>
+											<div className='text-zinc-600'>ID: {todo.id.slice(-8)}</div>
+										</div>
 									</div>
 								</div>
 							</div>
+						);
+					})}
+				</div>
+			</ScrollArea>
 
-							{/* Actions */}
-							<div className='flex flex-col gap-2 ml-4'>
-								{/* expand todo toggle button */}
-								<Button
-									aria-label='View full and close todo'
-									onClick={() => expandTodoToggle(todo.id)}
-									size='sm'
-									variant='default'
-									className='bg-zinc-700/60 hover:bg-zinc-600 text-zinc-200 border-zinc-600'>
-									{isExpanded ? <EyeClosed /> : <Eye />} {isExpanded ? 'Close' : 'View Full'}
-								</Button>
-								<Button
-									aria-label='Edit todo'
-									onClick={() => handleEditClick(todo)}
-									size='sm'
-									variant='default'
-									className='bg-zinc-700/60 hover:bg-zinc-600 text-zinc-200 border-zinc-600'>
-									<PenBox /> Edit
-								</Button>
+			{/* Modals */}
+			<Suspense
+				fallback={
+					<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
+						<div className='text-white'>Loading...</div>
+					</div>
+				}>
+				{isModalOpen && selectedTodo && (
+					<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
+						<DeleteModal
+							open={isModalOpen}
+							onClose={() => setIsModalOpen(false)}
+							onConfirm={handleDeleteConfirm}
+							itemName={selectedTodo?.title}
+						/>
+					</div>
+				)}
+			</Suspense>
 
-								<Button
-									aria-label='Delete todo'
-									onClick={() => handleDeleteClick(todo)}
-									size='sm'
-									variant='destructive'
-									className='hover:bg-red-700'>
-									<Trash /> Delete
-								</Button>
-							</div>
-						</div>
-					);
-				})}
-				{/* pagination buttons */}
-				<PaginationButtons
-					goToFirstPage={goToFirstPage}
-					currentPage={currentPage}
-					goToPreviousPage={goToPreviousPage}
-					totalPages={totalPages}
-					goToSpecificPage={goToSpecificPage}
-					goToNextPage={goToNextPage}
-					goToLastPage={goToLastPage}
-				/>
-				{/* Delete Modal */}
-				<Suspense fallback={<span>Loading Delete Modal...</span>}>
-					{isModalOpen && selectedTodo && (
-						<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
-							<DeleteModal
-								open={isModalOpen}
-								onClose={() => setIsModalOpen(false)}
-								onConfirm={handleDeleteConfirm}
-								itemName={selectedTodo?.title}
-							/>
-						</div>
-					)}
-				</Suspense>
-				{/* Edit Modal */}
-				<Suspense fallback={<span>Loading Edit Modal...</span>}>
-					{isEditModalOpen && editingTodo && (
-						<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
-							<EditModal
-								open={isEditModalOpen}
-								onClose={() => setIsEditModalOpen(false)}
-								todo={editingTodo}
-								onConfirm={handleEditConfirm}
-							/>
-						</div>
-					)}
-				</Suspense>
-			</div>
-		</ScrollArea>
+			<Suspense
+				fallback={
+					<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
+						<div className='text-white'>Loading...</div>
+					</div>
+				}>
+				{isEditModalOpen && editingTodo && (
+					<div className='fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm'>
+						<EditModal
+							open={isEditModalOpen}
+							onClose={() => setIsEditModalOpen(false)}
+							todo={editingTodo}
+							onConfirm={handleEditConfirm}
+						/>
+					</div>
+				)}
+			</Suspense>
+		</div>
 	);
 };
